# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A CRUD server for analyzing financial portfolios comprising stocks, bonds, and ETFs.",
        "title": "Portfolio API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/portfolios": {
            "post": {
                "description": "Create a new portfolio with optional memberships. Accepts JSON or multipart/form-data with a CSV file.",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolios"
                ],
                "summary": "Create a new portfolio",
                "parameters": [
                    {
                        "description": "Portfolio to create (JSON)",
                        "name": "portfolio",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePortfolioRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Portfolio metadata as JSON (multipart)",
                        "name": "metadata",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "CSV file with ticker,percentage_or_shares columns (multipart)",
                        "name": "memberships",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.PortfolioWithMemberships"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/portfolios/{id}": {
            "get": {
                "description": "Retrieve a portfolio and its memberships by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolios"
                ],
                "summary": "Get a portfolio by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portfolio ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PortfolioWithMemberships"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "UserID": []
                    }
                ],
                "description": "Update a portfolio's name and/or memberships. Accepts JSON or multipart/form-data with a CSV file.",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolios"
                ],
                "summary": "Update a portfolio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portfolio ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Portfolio updates (JSON)",
                        "name": "portfolio",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.UpdatePortfolioRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Portfolio metadata as JSON (multipart)",
                        "name": "metadata",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "CSV file with ticker,percentage_or_shares columns (multipart)",
                        "name": "memberships",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PortfolioWithMemberships"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserID": []
                    }
                ],
                "description": "Delete a portfolio by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolios"
                ],
                "summary": "Delete a portfolio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portfolio ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreatePortfolioRequest": {
            "type": "object",
            "required": [
                "name",
                "objective",
                "owner_id",
                "portfolio_type"
            ],
            "properties": {
                "created_at": {
                    "description": "CreatedAt is the date the portfolio began trading or was built.\nReflects when the strategy was initiated, not when it was imported into this system.\nAccepts \"YYYY-MM-DD\" or RFC3339. Defaults to the current timestamp if omitted.",
                    "type": "string",
                    "example": "2023-06-01"
                },
                "memberships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MembershipRequest"
                    }
                },
                "name": {
                    "type": "string"
                },
                "objective": {
                    "$ref": "#/definitions/models.Objective"
                },
                "owner_id": {
                    "type": "integer"
                },
                "portfolio_type": {
                    "$ref": "#/definitions/models.PortfolioType"
                },
                "snapshotted_at": {
                    "description": "SnapshottedAt records when the membership share counts were entered into this system.\nUsed to reverse stock splits that occurred between created_at and this date,\nso historical performance is computed with correct pre-split share counts.",
                    "type": "string",
                    "example": "2025-03-15"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.MembershipRequest": {
            "type": "object",
            "required": [
                "percentage_or_shares"
            ],
            "properties": {
                "percentage_or_shares": {
                    "type": "number"
                },
                "security_id": {
                    "type": "integer"
                },
                "ticker": {
                    "type": "string"
                }
            }
        },
...
        "models.Objective": {
            "type": "string",
            "enum": [
                "Aggressive Growth",
                "Growth",
                "Income Generation",
                "Capital Preservation",
                "Mixed Growth/Income"
            ],
            "x-enum-varnames": [
                "ObjectiveAggressiveGrowth",
                "ObjectiveGrowth",
                "ObjectiveIncomeGeneration",
                "ObjectiveCapitalPreservation",
                "ObjectiveMixedGrowthIncome"
            ]
        },
...
        "models.Portfolio": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "ended_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "objective": {
                    "$ref": "#/definitions/models.Objective"
                },
                "owner_id": {
                    "type": "integer"
                },
                "portfolio_type": {
                    "$ref": "#/definitions/models.PortfolioType"
                },
                "snapshotted_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.PortfolioMembership": {
            "type": "object",
            "properties": {
                "percentage_or_shares": {
                    "type": "number"
                },
                "portfolio_id": {
                    "type": "integer"
                },
                "security_id": {
                    "type": "integer"
                },
                "ticker": {
                    "type": "string"
                }
            }
        },
...
        "models.PortfolioType": {
            "type": "string",
            "enum": [
                "Ideal",
                "Active",
                "Historic"
            ],
            "x-enum-varnames": [
                "PortfolioTypeIdeal",
                "PortfolioTypeActive",
                "PortfolioTypeHistoric"
            ]
        },
        "models.PortfolioWithMemberships": {
            "type": "object",
            "properties": {
                "memberships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PortfolioMembership"
                    }
                },
                "portfolio": {
                    "$ref": "#/definitions/models.Portfolio"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Warning"
                    }
                }
            }
        },
...
        "models.UpdatePortfolioRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt is the date the portfolio began trading or was built.\nReflects when the strategy was initiated, not when it was imported into this system.\nAccepts \"YYYY-MM-DD\" or RFC3339. If omitted, the existing value is preserved.",
                    "type": "string",
                    "example": "2023-06-01"
                },
                "memberships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MembershipRequest"
                    }
                },
                "name": {
                    "type": "string"
                },
                "objective": {
                    "$ref": "#/definitions/models.Objective"
                },
                "portfolio_type": {
                    "$ref": "#/definitions/models.PortfolioType"
                },
                "snapshotted_at": {
                    "description": "SnapshottedAt records when the membership share counts were entered into this system.\nIf omitted, the existing value is preserved.",
                    "type": "string",
                    "example": "2025-03-15"
                }
            }
        },
...
        "models.Warning": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/models.WarningCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.WarningCode": {
            "type": "string",
            "enum": [
                "W1001",
                "W1002",
                "W1003",
                "W2001",
                "W3001",
                "W3002",
                "W4001",
                "W4002",
                "W4003",
                "W4004"
            ],
            "x-enum-comments": {
                "WarnBenchmarkDataUnavailable": "benchmark ticker missing or has no price data; Alpha/Beta set to zero",
                "WarnCashSubstituted": "pre-IPO period covered with synthetic cash prices; start date unchanged",
                "WarnETFSourceIncomplete": "source ETF data does not add up to 100%",
                "WarnExcessiveForwardFill": "too many securities needed forward-filling on some dates; those dates excluded",
                "WarnFuzzyMatchSubstituted": "dash-inserted ticker used in place of original (e.g. BRKB → BRK-B)",
                "WarnMissingPriceHistory": "one or more securities have no price history; affected dates excluded",
                "WarnPartialETFExpansion": "holdings scaled to 100% because resolved weights didn't sum to 1.0",
                "WarnProportionalReallocation": "pre-IPO period handled by redistributing weight among available members",
                "WarnStartDateAdjusted": "start date adjusted to security inception date",
                "WarnUnresolvedETFHolding": "individual unresolved holding (dropped from results)"
            },
            "x-enum-descriptions": [
                "individual unresolved holding (dropped from results)",
                "holdings scaled to 100% because resolved weights didn't sum to 1.0",
                "source ETF data does not add up to 100%",
                "dash-inserted ticker used in place of original (e.g. BRKB → BRK-B)",
                "one or more securities have no price history; affected dates excluded",
                "too many securities needed forward-filling on some dates; those dates excluded",
                "start date adjusted to security inception date",
                "benchmark ticker missing or has no price data; Alpha/Beta set to zero",
                "pre-IPO period covered with synthetic cash prices; start date unchanged",
                "pre-IPO period handled by redistributing weight among available members"
            ],
            "x-enum-varnames": [
                "WarnUnresolvedETFHolding",
                "WarnPartialETFExpansion",
                "WarnETFSourceIncomplete",
                "WarnFuzzyMatchSubstituted",
                "WarnMissingPriceHistory",
                "WarnExcessiveForwardFill",
                "WarnStartDateAdjusted",
                "WarnBenchmarkDataUnavailable",
                "WarnCashSubstituted",
                "WarnProportionalReallocation"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}